Move Acl::Node to its own header file
