+++ src/acl/Random.h
@@ -9,7 +9,7 @@
 #ifndef SQUID_ACL_RANDOM_H
 #define SQUID_ACL_RANDOM_H
 
-#include "acl/Acl.h"
+#include "acl/Node.h"
 
 class ACLRandom : public Acl::Node
 {
